+++ demos/components-slider-carousel.php
@@ -1,0 +1,425 @@
+<?php include 'include/head.php'; ?>
+<body>
+<div class="page">
+  <div class="container">
+
+    <div class="topic">
+      <h2 id=""><span>slider: </span>carousel</h2>
+    </div>
+    <div class="content">
+      <div class="slider">
+        <input type="radio" name="option-carousel" id="carousel-lg">
+        <input type="radio" name="option-carousel" id="carousel-md">
+        <input type="radio" name="option-carousel" id="carousel-sm">
+        <input type="radio" name="option-carousel" id="slide-by-page">
+        <input type="radio" name="option-carousel" id="slider-center">
+        <input type="radio" name="option-carousel" id="carousel-auto-play">
+        <input type="radio" name="option-carousel" id="carousel-auto-height" checked="">
+        <div class="slider-labels">
+          <div>
+            <div class="title">Options: </div>
+            <label for="carousel-lg">lg</label>
+            <label for="carousel-md">md</label>
+            <label for="carousel-sm">sm</label>
+            <label for="slide-by-page">slide-by-page</label>
+            <label for="slider-center">center</label>
+            <label for="carousel-auto-play">auto-play</label>
+            <label for="carousel-auto-height">auto-height</label>
+          </div>
+        </div>
+        <div class="carousel-a">
+          <input type="radio" name="carousel-a" id="carousel-a-1" checked="">
+          <input type="radio" name="carousel-a" id="carousel-a-2">
+          <input type="radio" name="carousel-a" id="carousel-a-3">
+          <input type="radio" name="carousel-a" id="carousel-a-4">
+          <input type="radio" name="carousel-a" id="carousel-a-5">
+          <input type="checkbox" name="carousel-a-autoplay" id="carousel-a-autoplay">
+          <div class="outer">
+            <ul class="inner">
+              <li>
+                <div class="item"><a href=""><img src="http://placehold.it/1100x619/97cddd/97cddd"></a>
+                  <div class="info"><h3>Cloud Dragon</h3> by <a href="http://voyager3.tumblr.com">Brendan Zabarauskas</a></div>
+                </div>
+              </li>
+              <li>
+                <div class="item"><a href=""><img src="http://placehold.it/1100x619/e3d189/e3d189"></a>
+                  <div class="info"><h3>Mountain Fort</h3> by <a href="http://voyager3.tumblr.com">Brendan Zabarauskas</a></div>
+                </div>
+              </li>
+              <li>
+                <div class="item"><a href=""><img src="http://placehold.it/1100x619/97cddd/97cddd"></a>
+                  <div class="info"><h3>Mountain Outpost</h3> by <a href="http://voyager3.tumblr.com">Brendan Zabarauskas</a></div>
+                </div>
+              </li>
+              <li>
+                <div class="item"><a href=""><img src="http://placehold.it/1100x619/b6cc84/b6cc84"></a>
+                  <div class="info"><h3>Cliffs</h3> by <a href="http://voyager3.tumblr.com">Brendan Zabarauskas</a></div>
+                </div>
+              </li>
+              <li>
+                <div class="item"><a href=""><img src="http://placehold.it/1100x619/eea7d4/eea7d4"></a>
+                  <div class="info"><h3>Hill Fort</h3> by <a href="http://voyager3.tumblr.com">Brendan Zabarauskas</a></div>
+                </div>
+              </li>
+            </ul>
+          </div>
+          <div class="controls">
+            <label for="carousel-a-1"><span class="prev"></span><span class="next"></span></label>
+            <label for="carousel-a-2"><span class="prev"></span><span class="next"></span></label>
+            <label for="carousel-a-3"><span class="prev"></span><span class="next"></span></label>
+            <label for="carousel-a-4"><span class="prev"></span><span class="next"></span></label>
+            <label for="carousel-a-5"><span class="prev"></span><span class="next"></span></label>
+          </div>
+          <div class="dots">
+            <label for="carousel-a-1"><span class="normal"></span><span class="active"></span></label>
+            <label for="carousel-a-2"><span class="normal"></span><span class="active"></span></label>
+            <label for="carousel-a-3"><span class="normal"></span><span class="active"></span></label>
+            <label for="carousel-a-4"><span class="normal"></span><span class="active"></span></label>
+            <label for="carousel-a-5"><span class="normal"></span><span class="active"></span></label>
+          </div>
+          <div class="autoplay"><label for="carousel-a-autoplay"><span></span></label></div>
+        </div>
+        <div class="carousel-b">
+          <input type="radio" name="carousel-b" id="carousel-b-1" checked="">
+          <input type="radio" name="carousel-b" id="carousel-b-2">
+          <input type="radio" name="carousel-b" id="carousel-b-3">
+          <input type="radio" name="carousel-b" id="carousel-b-4">
+          <input type="radio" name="carousel-b" id="carousel-b-5">
+          <input type="checkbox" name="carousel-b-autoplay" id="carousel-b-autoplay">
+          <div class="outer">
+            <ul class="inner">
+              <li>
+                <div class="item"><a href=""><img src="http://placehold.it/1100x619/97cddd/97cddd"></a>
+                  <div class="info"><h3>Cloud Dragon</h3> by <a href="http://voyager3.tumblr.com">Brendan Zabarauskas</a></div>
+                </div>
+              </li>
+              <li>
+                <div class="item"><a href=""><img src="http://placehold.it/1100x619/e3d189/e3d189"></a>
+                  <div class="info"><h3>Mountain Fort</h3> by <a href="http://voyager3.tumblr.com">Brendan Zabarauskas</a></div>
+                </div>
+              </li>
+              <li>
+                <div class="item"><a href=""><img src="http://placehold.it/1100x619/97cddd/97cddd"></a>
+                  <div class="info"><h3>Mountain Outpost</h3> by <a href="http://voyager3.tumblr.com">Brendan Zabarauskas</a></div>
+                </div>
+              </li>
+              <li>
+                <div class="item"><a href=""><img src="http://placehold.it/1100x619/b6cc84/b6cc84"></a>
+                  <div class="info"><h3>Cliffs</h3> by <a href="http://voyager3.tumblr.com">Brendan Zabarauskas</a></div>
+                </div>
+              </li>
+              <li>
+                <div class="item"><a href=""><img src="http://placehold.it/1100x619/eea7d4/eea7d4"></a>
+                  <div class="info"><h3>Hill Fort</h3> by <a href="http://voyager3.tumblr.com">Brendan Zabarauskas</a></div>
+                </div>
+              </li>
+            </ul>
+          </div>
+          <div class="controls">
+            <label for="carousel-b-1"><span class="prev"></span><span class="next"></span></label>
+            <label for="carousel-b-2"><span class="prev"></span><span class="next"></span></label>
+            <label for="carousel-b-3"><span class="prev"></span><span class="next"></span></label>
+            <label for="carousel-b-4"><span class="prev"></span><span class="next"></span></label>
+            <label for="carousel-b-5"><span class="prev"></span><span class="next"></span></label>
+          </div>
+          <div class="dots">
+            <label for="carousel-b-1"><span class="normal"></span><span class="active"></span></label>
+            <label for="carousel-b-2"><span class="normal"></span><span class="active"></span></label>
+            <label for="carousel-b-3"><span class="normal"></span><span class="active"></span></label>
+            <label for="carousel-b-4"><span class="normal"></span><span class="active"></span></label>
+            <label for="carousel-b-5"><span class="normal"></span><span class="active"></span></label>
+          </div>
+          <div class="autoplay"><label for="carousel-b-autoplay"><span></span></label></div>
+        </div>
+        <div class="carousel-c">
+          <input type="radio" name="carousel-c" id="carousel-c-1" checked="">
+          <input type="radio" name="carousel-c" id="carousel-c-2">
+          <input type="radio" name="carousel-c" id="carousel-c-3">
+          <input type="radio" name="carousel-c" id="carousel-c-4">
+          <input type="radio" name="carousel-c" id="carousel-c-5">
+          <input type="checkbox" name="carousel-c-autoplay" id="carousel-c-autoplay">
+          <div class="outer">
+            <ul class="inner">
+              <li>
+                <div class="item"><a href=""><img src="http://placehold.it/1100x619/97cddd/97cddd"></a>
+                  <div class="info"><h3>Cloud Dragon</h3> by <a href="http://voyager3.tumblr.com">Brendan Zabarauskas</a></div>
+                </div>
+              </li>
+              <li>
+                <div class="item"><a href=""><img src="http://placehold.it/1100x619/e3d189/e3d189"></a>
+                  <div class="info"><h3>Mountain Fort</h3> by <a href="http://voyager3.tumblr.com">Brendan Zabarauskas</a></div>
+                </div>
+              </li>
+              <li>
+                <div class="item"><a href=""><img src="http://placehold.it/1100x619/97cddd/97cddd"></a>
+                  <div class="info"><h3>Mountain Outpost</h3> by <a href="http://voyager3.tumblr.com">Brendan Zabarauskas</a></div>
+                </div>
+              </li>
+              <li>
+                <div class="item"><a href=""><img src="http://placehold.it/1100x619/b6cc84/b6cc84"></a>
+                  <div class="info"><h3>Cliffs</h3> by <a href="http://voyager3.tumblr.com">Brendan Zabarauskas</a></div>
+                </div>
+              </li>
+              <li>
+                <div class="item"><a href=""><img src="http://placehold.it/1100x619/eea7d4/eea7d4"></a>
+                  <div class="info"><h3>Hill Fort</h3> by <a href="http://voyager3.tumblr.com">Brendan Zabarauskas</a></div>
+                </div>
+              </li>
+            </ul>
+          </div>
+          <div class="controls">
+            <label for="carousel-c-1"><span class="prev"></span><span class="next"></span></label>
+            <label for="carousel-c-2"><span class="prev"></span><span class="next"></span></label>
+            <label for="carousel-c-3"><span class="prev"></span><span class="next"></span></label>
+            <label for="carousel-c-4"><span class="prev"></span><span class="next"></span></label>
+            <label for="carousel-c-5"><span class="prev"></span><span class="next"></span></label>
+          </div>
+          <div class="dots">
+            <label for="carousel-c-1"><span class="normal"></span><span class="active"></span></label>
+            <label for="carousel-c-2"><span class="normal"></span><span class="active"></span></label>
+            <label for="carousel-c-3"><span class="normal"></span><span class="active"></span></label>
+            <label for="carousel-c-4"><span class="normal"></span><span class="active"></span></label>
+            <label for="carousel-c-5"><span class="normal"></span><span class="active"></span></label>
+          </div>
+          <div class="autoplay"><label for="carousel-c-autoplay"><span></span></label></div>
+        </div>
+        <div class="carousel-d">
+          <input type="radio" name="carousel-d" id="carousel-d-1" checked="">
+          <input type="radio" name="carousel-d" id="carousel-d-2">
+          <input type="radio" name="carousel-d" id="carousel-d-3">
+          <input type="radio" name="carousel-d" id="carousel-d-4">
+          <input type="radio" name="carousel-d" id="carousel-d-5">
+          <input type="checkbox" name="carousel-d-autoplay" id="carousel-d-autoplay">
+          <div class="outer">
+            <ul class="inner">
+              <li>
+                <div class="item"><a href=""><img src="http://placehold.it/1100x619/97cddd/97cddd"></a>
+                  <div class="info"><h3>Cloud Dragon</h3> by <a href="http://voyager3.tumblr.com">Brendan Zabarauskas</a></div>
+                </div>
+              </li>
+              <li>
+                <div class="item"><a href=""><img src="http://placehold.it/1100x619/e3d189/e3d189"></a>
+                  <div class="info"><h3>Mountain Fort</h3> by <a href="http://voyager3.tumblr.com">Brendan Zabarauskas</a></div>
+                </div>
+              </li>
+              <li>
+                <div class="item"><a href=""><img src="http://placehold.it/1100x619/97cddd/97cddd"></a>
+                  <div class="info"><h3>Mountain Outpost</h3> by <a href="http://voyager3.tumblr.com">Brendan Zabarauskas</a></div>
+                </div>
+              </li>
+              <li>
+                <div class="item"><a href=""><img src="http://placehold.it/1100x619/b6cc84/b6cc84"></a>
+                  <div class="info"><h3>Cliffs</h3> by <a href="http://voyager3.tumblr.com">Brendan Zabarauskas</a></div>
+                </div>
+              </li>
+              <li>
+                <div class="item"><a href=""><img src="http://placehold.it/1100x619/eea7d4/eea7d4"></a>
+                  <div class="info"><h3>Hill Fort</h3> by <a href="http://voyager3.tumblr.com">Brendan Zabarauskas</a></div>
+                </div>
+              </li>
+            </ul>
+          </div>
+          <div class="controls">
+            <label for="carousel-d-1"><span class="prev"></span><span class="next"></span></label>
+            <label for="carousel-d-2"><span class="prev"></span><span class="next"></span></label>
+            <label for="carousel-d-3"><span class="prev"></span><span class="next"></span></label>
+            <label for="carousel-d-4"><span class="prev"></span><span class="next"></span></label>
+            <label for="carousel-d-5"><span class="prev"></span><span class="next"></span></label>
+          </div>
+          <div class="dots">
+            <label for="carousel-d-1"><span class="normal"></span><span class="active"></span></label>
+            <label for="carousel-d-2"><span class="normal"></span><span class="active"></span></label>
+            <label for="carousel-d-3"><span class="normal"></span><span class="active"></span></label>
+            <label for="carousel-d-4"><span class="normal"></span><span class="active"></span></label>
+            <label for="carousel-d-5"><span class="normal"></span><span class="active"></span></label>
+          </div>
+          <div class="autoplay"><label for="carousel-d-autoplay"><span></span></label></div>
+        </div>
+        <div class="carousel-e">
+          <input type="radio" name="carousel-e" id="carousel-e-1" checked="">
+          <input type="radio" name="carousel-e" id="carousel-e-2">
+          <input type="radio" name="carousel-e" id="carousel-e-3">
+          <input type="radio" name="carousel-e" id="carousel-e-4">
+          <input type="radio" name="carousel-e" id="carousel-e-5">
+          <input type="checkbox" name="carousel-e-autoplay" id="carousel-e-autoplay">
+          <div class="outer">
+            <ul class="inner">
+              <li>
+                <div class="item"><a href=""><img src="http://placehold.it/1100x619/97cddd/97cddd"></a>
+                  <div class="info"><h3>Cloud Dragon</h3> by <a href="http://voyager3.tumblr.com">Brendan Zabarauskas</a></div>
+                </div>
+              </li>
+              <li>
+                <div class="item"><a href=""><img src="http://placehold.it/1100x619/e3d189/e3d189"></a>
+                  <div class="info"><h3>Mountain Fort</h3> by <a href="http://voyager3.tumblr.com">Brendan Zabarauskas</a></div>
+                </div>
+              </li>
+              <li>
+                <div class="item"><a href=""><img src="http://placehold.it/1100x619/97cddd/97cddd"></a>
+                  <div class="info"><h3>Mountain Outpost</h3> by <a href="http://voyager3.tumblr.com">Brendan Zabarauskas</a></div>
+                </div>
+              </li>
+              <li>
+                <div class="item"><a href=""><img src="http://placehold.it/1100x619/b6cc84/b6cc84"></a>
+                  <div class="info"><h3>Cliffs</h3> by <a href="http://voyager3.tumblr.com">Brendan Zabarauskas</a></div>
+                </div>
+              </li>
+              <li>
+                <div class="item"><a href=""><img src="http://placehold.it/1100x619/eea7d4/eea7d4"></a>
+                  <div class="info"><h3>Hill Fort</h3> by <a href="http://voyager3.tumblr.com">Brendan Zabarauskas</a></div>
+                </div>
+              </li>
+            </ul>
+          </div>
+          <div class="controls">
+            <label for="carousel-e-1"><span class="prev"></span><span class="next"></span></label>
+            <label for="carousel-e-2"><span class="prev"></span><span class="next"></span></label>
+            <label for="carousel-e-3"><span class="prev"></span><span class="next"></span></label>
+            <label for="carousel-e-4"><span class="prev"></span><span class="next"></span></label>
+            <label for="carousel-e-5"><span class="prev"></span><span class="next"></span></label>
+          </div>
+          <div class="dots">
+            <label for="carousel-e-1"><span class="normal"></span><span class="active"></span></label>
+            <label for="carousel-e-2"><span class="normal"></span><span class="active"></span></label>
+            <label for="carousel-e-3"><span class="normal"></span><span class="active"></span></label>
+            <label for="carousel-e-4"><span class="normal"></span><span class="active"></span></label>
+            <label for="carousel-e-5"><span class="normal"></span><span class="active"></span></label>
+          </div>
+          <div class="autoplay"><label for="carousel-e-autoplay"><span></span></label></div>
+        </div>
+        <div class="carousel-f">
+          <input type="radio" name="carousel-f" id="carousel-f-1" checked="">
+          <input type="radio" name="carousel-f" id="carousel-f-2">
+          <input type="radio" name="carousel-f" id="carousel-f-3">
+          <input type="radio" name="carousel-f" id="carousel-f-4">
+          <input type="radio" name="carousel-f" id="carousel-f-5">
+          <input type="checkbox" name="carousel-f-autoplay" id="carousel-f-autoplay" checked="">
+          <div class="outer">
+            <ul class="inner">
+              <li>
+                <div class="item"><a href=""><img src="http://placehold.it/1100x619/97cddd/97cddd"></a>
+                  <div class="info"><h3>Cloud Dragon</h3> by <a href="http://voyager3.tumblr.com">Brendan Zabarauskas</a></div>
+                </div>
+              </li>
+              <li>
+                <div class="item"><a href=""><img src="http://placehold.it/1100x619/e3d189/e3d189"></a>
+                  <div class="info"><h3>Mountain Fort</h3> by <a href="http://voyager3.tumblr.com">Brendan Zabarauskas</a></div>
+                </div>
+              </li>
+              <li>
+                <div class="item"><a href=""><img src="http://placehold.it/1100x619/97cddd/97cddd"></a>
+                  <div class="info"><h3>Mountain Outpost</h3> by <a href="http://voyager3.tumblr.com">Brendan Zabarauskas</a></div>
+                </div>
+              </li>
+              <li>
+                <div class="item"><a href=""><img src="http://placehold.it/1100x619/b6cc84/b6cc84"></a>
+                  <div class="info"><h3>Cliffs</h3> by <a href="http://voyager3.tumblr.com">Brendan Zabarauskas</a></div>
+                </div>
+              </li>
+              <li>
+                <div class="item"><a href=""><img src="http://placehold.it/1100x619/eea7d4/eea7d4"></a>
+                  <div class="info"><h3>Hill Fort</h3> by <a href="http://voyager3.tumblr.com">Brendan Zabarauskas</a></div>
+                </div>
+              </li>
+            </ul>
+          </div>
+          <div class="controls">
+            <label for="carousel-f-1"><span class="prev"></span><span class="next"></span></label>
+            <label for="carousel-f-2"><span class="prev"></span><span class="next"></span></label>
+            <label for="carousel-f-3"><span class="prev"></span><span class="next"></span></label>
+            <label for="carousel-f-4"><span class="prev"></span><span class="next"></span></label>
+            <label for="carousel-f-5"><span class="prev"></span><span class="next"></span></label>
+          </div>
+          <div class="dots">
+            <label for="carousel-f-1"><span class="normal"></span><span class="active"></span></label>
+            <label for="carousel-f-2"><span class="normal"></span><span class="active"></span></label>
+            <label for="carousel-f-3"><span class="normal"></span><span class="active"></span></label>
+            <label for="carousel-f-4"><span class="normal"></span><span class="active"></span></label>
+            <label for="carousel-f-5"><span class="normal"></span><span class="active"></span></label>
+          </div>
+          <div class="autoplay"><label for="carousel-f-autoplay"><span></span></label></div>
+        </div>
+        <div class="carousel-g">
+          <input type="radio" name="carousel-g" id="carousel-g-1" checked="">
+          <input type="radio" name="carousel-g" id="carousel-g-2">
+          <input type="radio" name="carousel-g" id="carousel-g-3">
+          <input type="radio" name="carousel-g" id="carousel-g-4">
+          <input type="radio" name="carousel-g" id="carousel-g-5">
+          <input type="checkbox" name="carousel-g-autoplay" id="carousel-g-autoplay" checked="">
+          <div class="outer" autoheight-carousel>
+            <ul class="inner">
+              <li>
+                <div class="item"><a href=""><img src="http://placehold.it/1100x919/97cddd/97cddd"></a>
+                  <div class="info"><h3>Cloud Dragon</h3> by <a href="http://voyager3.tumblr.com">Brendan Zabarauskas</a></div>
+                </div>
+              </li>
+              <li>
+                <div class="item"><a href=""><img src="http://placehold.it/1100x719/e3d189/e3d189"></a>
+                  <div class="info"><h3>Mountain Fort</h3> by <a href="http://voyager3.tumblr.com">Brendan Zabarauskas</a></div>
+                </div>
+              </li>
+              <li>
+                <div class="item"><a href=""><img src="http://placehold.it/1100x619/97cddd/97cddd"></a>
+                  <div class="info"><h3>Mountain Outpost</h3> by <a href="http://voyager3.tumblr.com">Brendan Zabarauskas</a></div>
+                </div>
+              </li>
+              <li>
+                <div class="item"><a href=""><img src="http://placehold.it/1100x519/b6cc84/b6cc84"></a>
+                  <div class="info"><h3>Cliffs</h3> by <a href="http://voyager3.tumblr.com">Brendan Zabarauskas</a></div>
+                </div>
+              </li>
+              <li>
+                <div class="item"><a href=""><img src="http://placehold.it/1100x619/eea7d4/eea7d4"></a>
+                  <div class="info"><h3>Hill Fort</h3> by <a href="http://voyager3.tumblr.com">Brendan Zabarauskas</a></div>
+                </div>
+              </li>
+            </ul>
+          </div>
+          <div class="controls">
+            <label for="carousel-g-1"><span class="prev"></span><span class="next"></span></label>
+            <label for="carousel-g-2"><span class="prev"></span><span class="next"></span></label>
+            <label for="carousel-g-3"><span class="prev"></span><span class="next"></span></label>
+            <label for="carousel-g-4"><span class="prev"></span><span class="next"></span></label>
+            <label for="carousel-g-5"><span class="prev"></span><span class="next"></span></label>
+          </div>
+          <div class="dots">
+            <label for="carousel-g-1"><span class="normal"></span><span class="active"></span></label>
+            <label for="carousel-g-2"><span class="normal"></span><span class="active"></span></label>
+            <label for="carousel-g-3"><span class="normal"></span><span class="active"></span></label>
+            <label for="carousel-g-4"><span class="normal"></span><span class="active"></span></label>
+            <label for="carousel-g-5"><span class="normal"></span><span class="active"></span></label>
+          </div>
+          <div class="autoplay"><label for="carousel-g-autoplay"><span></span></label></div>
+        </div>
+      </div>
+      <div>
+        <h4>SCSS</h4>
+        <pre class="language-scss"><code>
+.carousel-a { @include slider-carousel(5 by 1 default); }
+.carousel-b { @include slider-carousel(5 by 1 default); }
+.carousel-c { @include slider-carousel(5 by 1 default); }
+.carousel-d { @include slider-carousel(5 by 2 bypage default); }
+.carousel-e { @include slider-carousel(5 by 2 center default); }
+.carousel-f { @include slider-carousel(5 by 2 autoplay default); }
+.carousel-g { @include slider-carousel(5 by 1 default); }
+@include breakpoint-mi(em(700)) {
+  .carousel-a { @include slider-carousel(5 by 2); }
+  .carousel-b { @include slider-carousel(5 by 2); }
+  .carousel-d { @include slider-carousel(5 by 3 bypage); }
+  .carousel-g { @include slider-carousel(5 by 2); }
+}
+@include breakpoint-mi(em(1000)) {
+  .carousel-a { @include slider-carousel(5 by 3); }
+  .carousel-g { @include slider-carousel(5 by 3); }
+}
+        </code></pre>
+      </div>
+    </div>
+    
+    <?php include "include/more-demos.php" ?>
+  </div>
+  <?php include "include/site-footer.php" ?>
+</div>
+</body>
+</html>
+