+++ demos/components-slider-gallery.php
@@ -1,0 +1,218 @@
+<?php include 'include/head.php'; ?>
+<body>
+<div class="page">
+  <div class="container">
+
+    <div class="topic">
+      <h2 id=""><span>slider: </span>gallery</h2>
+      <div class="content">
+        <div class="content-main">
+          <div class="slider">
+            <input type="radio" name="option-gallery" id="gallery-auto-play">
+            <input type="radio" name="option-gallery" id="gallery-auto-height">
+            <input type="radio" name="option-gallery" id="gallery-customise" checked="">
+            <div class="slider-labels">
+              <div>
+                <div class="title">Options: </div>
+                <label for="gallery-auto-play">auto-play</label>
+                <label for="gallery-auto-height">auto-height</label>
+                <label for="gallery-customise">customise</label>
+              </div>
+            </div>
+            <div class="gallery-a">
+              <input type="radio" name="gallery-a" id="gallery-a-1" checked="">
+              <input type="radio" name="gallery-a" id="gallery-a-2">
+              <input type="radio" name="gallery-a" id="gallery-a-3">
+              <input type="radio" name="gallery-a" id="gallery-a-4">
+              <input type="radio" name="gallery-a" id="gallery-a-5">
+              <input type="checkbox" name="gallery-a-autoplay" id="gallery-a-autoplay" checked="">
+              <div class="outer">
+                <ul class="inner">
+                  <li>
+                    <div class="item"><a href=""><img src="http://placehold.it/1100x619/97cddd/97cddd"></a>
+                      <div class="info"><h3>Cloud Dragon</h3> by <a href="http://voyager3.tumblr.com">Brendan Zabarauskas</a></div>
+                    </div>
+                  </li>
+                  <li>
+                    <div class="item"><a href=""><img src="http://placehold.it/1100x619/e3d189/e3d189"></a>
+                      <div class="info"><h3>Mountain Fort</h3> by <a href="http://voyager3.tumblr.com">Brendan Zabarauskas</a></div>
+                    </div>
+                  </li>
+                  <li>
+                    <div class="item"><a href=""><img src="http://placehold.it/1100x619/97cddd/97cddd"></a>
+                      <div class="info"><h3>Mountain Outpost</h3> by <a href="http://voyager3.tumblr.com">Brendan Zabarauskas</a></div>
+                    </div>
+                  </li>
+                  <li>
+                    <div class="item"><a href=""><img src="http://placehold.it/1100x619/b6cc84/b6cc84"></a>
+                      <div class="info"><h3>Cliffs</h3> by <a href="http://voyager3.tumblr.com">Brendan Zabarauskas</a></div>
+                    </div>
+                  </li>
+                  <li>
+                    <div class="item"><a href=""><img src="http://placehold.it/1100x619/eea7d4/eea7d4"></a>
+                      <div class="info"><h3>Hill Fort</h3> by <a href="http://voyager3.tumblr.com">Brendan Zabarauskas</a></div>
+                    </div>
+                  </li>
+                </ul>
+              </div>
+              <div class="controls">
+                <label for="gallery-a-1"><span class="prev"></span><span class="next"></span></label>
+                <label for="gallery-a-2"><span class="prev"></span><span class="next"></span></label>
+                <label for="gallery-a-3"><span class="prev"></span><span class="next"></span></label>
+                <label for="gallery-a-4"><span class="prev"></span><span class="next"></span></label>
+                <label for="gallery-a-5"><span class="prev"></span><span class="next"></span></label>
+              </div>
+              <div class="dots">
+                <label for="gallery-a-1"><span class="normal"></span><span class="active"></span></label>
+                <label for="gallery-a-2"><span class="normal"></span><span class="active"></span></label>
+                <label for="gallery-a-3"><span class="normal"></span><span class="active"></span></label>
+                <label for="gallery-a-4"><span class="normal"></span><span class="active"></span></label>
+                <label for="gallery-a-5"><span class="normal"></span><span class="active"></span></label>
+              </div>
+              <div class="autoplay"><label for="gallery-a-autoplay"><span></span></label></div>
+            </div>
+            <div class="gallery-b">
+              <input type="radio" name="gallery-b" id="gallery-b-1" checked="">
+              <input type="radio" name="gallery-b" id="gallery-b-2">
+              <input type="radio" name="gallery-b" id="gallery-b-3">
+              <input type="radio" name="gallery-b" id="gallery-b-4">
+              <input type="radio" name="gallery-b" id="gallery-b-5">
+              <input type="checkbox" name="gallery-b-autoplay" id="gallery-b-autoplay" checked="">
+              <div class="outer" autoheight-gallery>
+                <ul class="inner">
+                  <li>
+                    <div class="item"><a href=""><img src="http://placehold.it/1100x919/97cddd/97cddd"></a>
+                      <div class="info"><h3>Cloud Dragon</h3> by <a href="http://voyager3.tumblr.com">Brendan Zabarauskas</a></div>
+                    </div>
+                  </li>
+                  <li>
+                    <div class="item"><a href=""><img src="http://placehold.it/1100x619/e3d189/e3d189"></a>
+                      <div class="info"><h3>Mountain Fort</h3> by <a href="http://voyager3.tumblr.com">Brendan Zabarauskas</a></div>
+                    </div>
+                  </li>
+                  <li>
+                    <div class="item"><a href=""><img src="http://placehold.it/1100x819/97cddd/97cddd"></a>
+                      <div class="info"><h3>Mountain Outpost</h3> by <a href="http://voyager3.tumblr.com">Brendan Zabarauskas</a></div>
+                    </div>
+                  </li>
+                  <li>
+                    <div class="item"><a href=""><img src="http://placehold.it/1100x619/b6cc84/b6cc84"></a>
+                      <div class="info"><h3>Cliffs</h3> by <a href="http://voyager3.tumblr.com">Brendan Zabarauskas</a></div>
+                    </div>
+                  </li>
+                  <li>
+                    <div class="item"><a href=""><img src="http://placehold.it/1100x419/eea7d4/eea7d4"></a>
+                      <div class="info"><h3>Hill Fort</h3> by <a href="http://voyager3.tumblr.com">Brendan Zabarauskas</a></div>
+                    </div>
+                  </li>
+                </ul>
+              </div>
+              <div class="controls">
+                <label for="gallery-b-1"><span class="prev"></span><span class="next"></span></label>
+                <label for="gallery-b-2"><span class="prev"></span><span class="next"></span></label>
+                <label for="gallery-b-3"><span class="prev"></span><span class="next"></span></label>
+                <label for="gallery-b-4"><span class="prev"></span><span class="next"></span></label>
+                <label for="gallery-b-5"><span class="prev"></span><span class="next"></span></label>
+              </div>
+              <div class="dots">
+                <label for="gallery-b-1"><span class="normal"></span><span class="active"></span></label>
+                <label for="gallery-b-2"><span class="normal"></span><span class="active"></span></label>
+                <label for="gallery-b-3"><span class="normal"></span><span class="active"></span></label>
+                <label for="gallery-b-4"><span class="normal"></span><span class="active"></span></label>
+                <label for="gallery-b-5"><span class="normal"></span><span class="active"></span></label>
+              </div>
+              <div class="autoplay"><label for="gallery-b-autoplay"><span></span></label>
+            </div>
+            </div>
+            <div class="gallery-c">
+              <input type="radio" name="gallery-c" id="gallery-c-1" checked="">
+              <input type="radio" name="gallery-c" id="gallery-c-2">
+              <input type="radio" name="gallery-c" id="gallery-c-3">
+              <input type="radio" name="gallery-c" id="gallery-c-4">
+              <input type="radio" name="gallery-c" id="gallery-c-5">
+              <input type="checkbox" name="gallery-c-autoplay" id="gallery-c-autoplay" checked="">
+              <div class="outer">
+                <ul class="inner">
+                  <li>
+                    <div class="item"><a href=""><img src="http://placehold.it/1100x619/97cddd/97cddd"></a>
+                      <div class="info"><h3>Cloud Dragon</h3> by <a href="http://voyager3.tumblr.com">Brendan Zabarauskas</a></div>
+                    </div>
+                  </li>
+                  <li>
+                    <div class="item"><a href=""><img src="http://placehold.it/1100x619/e3d189/e3d189"></a>
+                      <div class="info"><h3>Mountain Fort</h3> by <a href="http://voyager3.tumblr.com">Brendan Zabarauskas</a></div>
+                    </div>
+                  </li>
+                  <li>
+                    <div class="item"><a href=""><img src="http://placehold.it/1100x619/97cddd/97cddd"></a>
+                      <div class="info"><h3>Mountain Outpost</h3> by <a href="http://voyager3.tumblr.com">Brendan Zabarauskas</a></div>
+                    </div>
+                  </li>
+                  <li>
+                    <div class="item"><a href=""><img src="http://placehold.it/1100x619/b6cc84/b6cc84"></a>
+                      <div class="info"><h3>Cliffs</h3> by <a href="http://voyager3.tumblr.com">Brendan Zabarauskas</a></div>
+                    </div>
+                  </li>
+                  <li>
+                    <div class="item"><a href=""><img src="http://placehold.it/1100x619/eea7d4/eea7d4"></a>
+                      <div class="info"><h3>Hill Fort</h3> by <a href="http://voyager3.tumblr.com">Brendan Zabarauskas</a></div>
+                    </div>
+                  </li>
+                </ul>
+              </div>
+              <div class="controls">
+                <label for="gallery-c-1"><span class="prev"></span><span class="next"></span></label>
+                <label for="gallery-c-2"><span class="prev"></span><span class="next"></span></label>
+                <label for="gallery-c-3"><span class="prev"></span><span class="next"></span></label>
+                <label for="gallery-c-4"><span class="prev"></span><span class="next"></span></label>
+                <label for="gallery-c-5"><span class="prev"></span><span class="next"></span></label>
+              </div>
+              <div class="dots">
+                <label for="gallery-c-1"><span class="normal"></span><span class="active"></span></label>
+                <label for="gallery-c-2"><span class="normal"></span><span class="active"></span></label>
+                <label for="gallery-c-3"><span class="normal"></span><span class="active"></span></label>
+                <label for="gallery-c-4"><span class="normal"></span><span class="active"></span></label>
+                <label for="gallery-c-5"><span class="normal"></span><span class="active"></span></label>
+              </div>
+              <div class="autoplay"><label for="gallery-c-autoplay"><span></span></label>
+            </div>
+          </div>
+          </div>
+        </div>
+        <div class="content-aside">
+          <pre class="language-scss"><code>
+.gallery-a { @include slider-gallery(5 autoplay default); }
+.gallery-b { @include slider-gallery(5 default); }
+.gallery-c { @include slider-gallery(5 default); }
+
+.gallery-c {
+  // setting item
+  .outer { overflow: visible; }
+  li { @include transform(scale(1.1)); }
+  @for $i from 1 through 5 {
+    #gallery-c-#{$i}:checked ~ .outer li:nth-child(#{$i}) { @include transform(scale(1)); }
+  }
+  // setting .info
+  .info { opacity: 0; margin-left: 0;
+    -webkit-transition: all 1s $global-bezier 0.5s;
+    transition: all 1s $global-bezier 0.5s;
+   }
+  @for $i from 1 through 5 {
+    #gallery-c-#{$i}:checked ~ .outer li:nth-child(#{$i}) .info {
+      opacity: 1;
+      margin-left: 20px;
+    }
+  }
+}
+          </code></pre>
+        </div>
+      </div>
+    </div>
+    
+    <?php include "include/more-demos.php" ?>
+  </div>
+  <?php include "include/site-footer.php" ?>
+</div>
+</body>
+</html>
+