+++ src/js/kit.js
@@ -7,3 +7,4 @@
 
 // *** components ***
 // @codekit-prepend "components/offcanvas.js"
+// @codekit-prepend "components/sticky.js"